refactor: split SettingsView.swift into separate files
Split large SettingsView.swift (1475 lines) into three files to fix
SwiftLint file_length violation (max 1000 lines):

- SettingsView.swift: Core settings, playback, and shared components (844 lines)
- CertificateSettings.swift: Certificate/security settings views
- ParentalControls.swift: Parental controls, PIN, and content rating
